requirements.txt
asgiref==3.7.2
boto==2.49.0
boto3==1.26.161
botocore==1.29.161
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-csp==3.7
django-environ==0.10.0
django-extensions==3.2.3
django-on-heroku==1.1.2
django-recaptcha==3.0.0
django-storages==1.13.2
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
whitenoise==6.5.0
